# api/openapi.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  version: 1.0.0
  description: >
    Task management API for the "GitHub Copilot for Automation Engineers"
    course. Used in Module 4 Lab 16 as the source of truth for generating
    pytest API tests. Open this file in VS Code before prompting Copilot.

servers:
  - url: http://localhost:3001
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  /tasks:
    get:
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, completed, overdue, archived] }
        - name: priority
          in: query
          schema: { type: string, enum: [low, medium, high] }
        - name: search
          in: query
          schema: { type: string, maxLength: 100 }
        - name: user_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /tasks/{id}:
    get:
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404':
          description: Task not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400':
          description: Validation error
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, nullable: true, maxLength: 2000 }
        due_date: { type: string, format: date, nullable: true }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [active, completed, overdue, archived] }
        user_id: { type: integer, nullable: true }
        created_at: { type: string }
        updated_at: { type: string }
    TaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, nullable: true, maxLength: 2000 }
        due_date: { type: string, format: date, nullable: true }
        priority: { type: string, enum: [low, medium, high], default: medium }
        status: { type: string, enum: [active, completed, overdue, archived], default: active }
        user_id: { type: integer, nullable: true }
    Error:
      type: object
      properties:
        error: { type: string }
        errors:
          type: array
          items: { type: string }